added mock structured
